More performance
Up to a 2x increase in speed across the board.

Except that `u8`/`i8` are slower. ~19ns per iteration rather than ~10ns.

I'm sure this has something to do with compiler internals and pointer types.

`u8`/`i8` are plenty fast, and the bump for everything else is nice enough that I'll take it.